Granny chase, level structure, pickups
Granny will chase you if she is in the same room as you. She won't chase you fi you aren't or if you are hidden.

Items cna be picked up, their type set via an enum (key or walkietalkie), can be expanded to include batteries if that becomes a thing.